Move Player class to its own file
